local-tools/requirements.txt
numpy==1.25.2
openai_whisper==20230918
sounddevice==0.4.6
sox==1.4.1
